requirements.txt
python-dotenv
boto3
requests
pandas
psycopg2-binary>=2.9
ultralytics
pyarrow
fastparquet
tqdm
geopandas
SQLAlchemy